requirements.txt
python-dotenv==0.19.2
av==8.0.3
aiortc==1.2.1
fastapi==0.70.0
uvicorn==0.15.0
opencv-python-headless==4.5.4.60
pyvirtualcam==0.8.0
numba==0.54.1
numpy==1.20.3
Jinja2==3.0.3